embeds :smile:

+++ -/commands/ping.js
@@ -1,9 +1,17 @@
+const { MessageEmbed } = require("discord.js");
+const ee = require("../embed.json");
 module.exports = {
   name: 'ping',
   aliases: [],
 
   execute(message,args, cmd, client, discord) {
+    if(!message.guild) return;
     if (message.deletable) message.delete();
-      message.channel.send(` **${client.ws.ping}ms** `);
+      message.channel.send(new MessageEmbed()
+        .setColor(ee.color)
+        .setAuthor(`${client.ws.ping}ms`)
+        .setTimestamp()
+        .setFooter(ee.footertext)
+      );
   },
 };
